+++ src/Dimmr/Interop/NativeMethods.cs
@@ -16,6 +16,7 @@ internal static class NativeMethods
 
     // SetWindowPos flags.
     public static readonly IntPtr HWND_TOPMOST = new(-1);
+    public static readonly IntPtr HWND_NOTOPMOST = new(-2);
     public const uint SWP_NOSIZE = 0x0001;
     public const uint SWP_NOMOVE = 0x0002;
     public const uint SWP_NOACTIVATE = 0x0010;